requirements.txt
flask
flask-cors
flask-jwt-extended
python-dotenv

pymongo
requests

pinecone-client
groq
ollama

pdfplumber
cloudinary
